requirements.txt
# LLM & RAG - Simplified
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0

# Environment and utilities
python-dotenv>=1.0.0
click>=8.0.0
rich>=12.0.0
pyyaml>=6.0
gitpython>=3.1.0

# Vector Operations - Minimal
numpy>=1.21.0
tiktoken>=0.5.0